signing-app/pyproject.toml
[project]
name = "signing-app"
version = "0.1.0"
description = "Luka Pona Learning App"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "imageio>=2.37.0",
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
]
